clean up! The major change is the relocation of all SIMD support functions into a dedicated internal header.
The "SIMD simulation" has been removed, and the header does not define any functions if the machine does not at least support SSE2.

MMX support might be added later, but it is not a priority. This would be somewhat equivalent to the previous attempt at "SIMD simulation".

I also attempted to standardize the coding style. The early development was a bit chaotic, and the sporadic nature of later development caused some small inconsistencies in the writing style. So I tried to harmonize everything, and I will probably write a document on the coding style to maintain soon.

Prefixes for global variables have been added, specifically. A capital "G" for global variables, and "GC" for global constants.

The "pfInternal_" prefix has been changed in favor of a simpler "pfi" prefix, which may be slightly less readable at first glance, but I didn’t like the old one... I prefer to be concise.

Other minor corrections were also made during this work.